Use corners only for global homography
Using only the corners to calculate the global homography improves the
visualization of the chessboard, because we are using only the points
on the chessboard, that are relevant.

We also removed the dead code where we find the extrinsic parameters
of the camera, when we have a camera configuration file available.
This is unnecessary, because we do not use these extrinsic parameters
in the stereo calibration.