Make a camera to camera calibration system
A camera to camera calibration system is a simpler model of the
projector to camera calibration that we are attempting to solve.
It is easier to debug this system due to the abundance of resources
pertaining to the topic. Debugging this simple system may perhaps
help us pinpoint the source of the problem in our more complex
calibration system.

The code is structured almost identical to the projector to camera
calibration system. There is a script that captures images and stores
them to a directory and a calibration script that uses these images
to calibrate the stereo system. A recording script will be written
in the future to accommodate fast, systematic saving of each
experiment.