Implement a GUI for camera image capturing
A GUI greatly simplifies the image capturing process, because it
eliminates the need to toggle to the builtin Camera app to check
the view of the camera. With a GUI displaying the views of both
cameras, we can easily adjust the chessboard pattern to fit in both
views and afterwards, simply click the "Capture" button to save the
captured images to a directory.