Group captures and visualizations
Organizing the capture_* directories under a single capture directory
allows for a cleaner file structure, the effects of which become
more pronounced as the number of captures grows larger. The same can
be said of visualization files, because the calibration process
generates three files for each capture or orientation.

The naming of the capture directory in camera2camera has also been
modified to match the naming in the projector-camera calibration.